Added the GOPortFactory and GOSoundPortFactory classes and extracted a part from GOSoundPort there https://github.com/GrandOrgue/grandorgue/issues/703
