+++ src/grandorgue/settings/GOPortFactory.cpp
@@ -1,0 +1,9 @@
+/*
+* Copyright 2006 Milan Digital Audio LLC
+* Copyright 2009-2021 GrandOrgue contributors (see AUTHORS)
+* License GPL-2.0 or later (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
+*/
+
+#include "GOPortFactory.h"
+
+const std::vector<wxString> GOPortFactory::c_NoApis;
