+++ src/grandorgue/settings/GOSettings.cpp
@@ -4,27 +4,33 @@
 * License GPL-2.0 or later (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
 */
 
-#include "settings/GOSettings.h"
+#include "GOSettings.h"
 
-#include "sound/GOSoundDefs.h"
+#include <wx/filename.h>
+#include <wx/log.h>
+#include <wx/stdpaths.h>
+#include <wx/thread.h>
+
 #include "archive/GOArchiveFile.h"
 #include "config/GOConfigFileWriter.h"
 #include "config/GOConfigFileReader.h"
 #include "config/GOConfigReader.h"
 #include "config/GOConfigReaderDB.h"
 #include "config/GOConfigWriter.h"
+#include "settings/GOSettingEnum.cpp"
+#include "settings/GOSettingNumber.cpp"
+#include "sound/GOSoundDefs.h"
+#include "sound/ports/GOSoundPort.h"
+#include "sound/ports/GOSoundPortFactory.h"
 #include "go_limits.h"
 #include "GOMemoryPool.h"
 #include "GOOrgan.h"
 #include "GOPath.h"
-#include "settings/GOSettingEnum.cpp"
-#include "settings/GOSettingNumber.cpp"
 #include "GOStdPath.h"
-#include "sound/ports/GOSoundPort.h"
-#include <wx/filename.h>
-#include <wx/log.h>
-#include <wx/stdpaths.h>
-#include <wx/thread.h>
+#include "GOPortFactory.h"
+
+static const wxString SOUND_PORTS = wxT("SoundPorts");
+static const wxString ENABLED = wxT(".Enabled");
 
 const GOMidiSetting GOSettings:: m_MIDISettings[] = {
 	{ MIDI_RECV_MANUAL, 1, wxTRANSLATE("Manuals"), wxTRANSLATE("Pedal") },
@@ -157,6 +163,42 @@ GOSettings::~GOSettings()
 	Flush();
 }
 
+void load_ports_config(
+  GOConfigReader& cfg, const wxString& groupName, const GOPortFactory& factory,
+  GOPortsConfig& portsConfig
+)
+{
+  portsConfig.Clear();
+  for (const wxString &portName: factory.GetPortNames())
+  {
+    const bool isPortEnabled
+      = cfg.ReadBoolean(CMBSetting, groupName, portName + ENABLED, false, true);
+    const wxString prefix = portName + ".";
+
+    portsConfig.SetConfigEnabled(portName, isPortEnabled);
+
+    for (const wxString &apiName: factory.GetPortApiNames(portName))
+      portsConfig.SetConfigEnabled(
+	portName, apiName,
+	cfg.ReadBoolean(CMBSetting, groupName, prefix + apiName + ENABLED, false, true)
+      );
+  }
+}
+
+void save_ports_config(
+  GOConfigWriter& cfg, const wxString& groupName, const GOPortFactory& factory,
+  const GOPortsConfig& portsConfig
+)
+{
+  for (const wxString &portName: factory.GetPortNames())
+  {
+    cfg.WriteBoolean(groupName, portName + ENABLED, portsConfig.IsConfigEnabled(portName));
+    const wxString prefix = portName + ".";
+    for (const wxString &apiName: factory.GetPortApiNames(portName))
+      cfg.WriteBoolean(groupName, prefix + apiName + ENABLED, portsConfig.IsConfigEnabled(portName, apiName));
+  }
+}
+
 void GOSettings::Load()
 {
 	GOConfigFileReader cfg_file;
@@ -193,21 +235,12 @@ void GOSettings::Load()
 			m_AudioGroups.push_back(cfg.ReadString(CMBSetting, wxT("AudioGroups"), wxString::Format(wxT("Name%03d"), i + 1), false, wxString::Format(_("Audio group %d"), i + 1)));
 		if (!m_AudioGroups.size())
 			m_AudioGroups.push_back(_("Default audio group"));
-		
-		const wxString SOUND_PORTS = wxT("SoundPorts");
-		const wxString ENABLED = wxT(".Enabled");
-		m_PortsConfig.Clear();
-		for (const wxString &portName: GOSoundPort::getPortNames())
-		{
-		  const bool isPortEnabled = cfg.ReadBoolean(CMBSetting, SOUND_PORTS, portName + ENABLED, false, true);
-		  const wxString prefix = portName + ".";
 
-		  m_PortsConfig.SetConfigEnabled(portName, isPortEnabled);
+		load_ports_config(
+		  cfg, SOUND_PORTS, GOSoundPortFactory::getInstance(),
+		  m_SoundPortsConfig
+		);
 
-		  for (const wxString &apiName: GOSoundPort::getApiNames(portName))
-		    m_PortsConfig.SetConfigEnabled(portName, apiName, cfg.ReadBoolean(CMBSetting, SOUND_PORTS, prefix + apiName + ENABLED, false, true));
-		}
-		
 		m_AudioDeviceConfig.clear();
 		count = cfg.ReadInteger(CMBSetting, wxT("AudioDevices"), wxT("Count"), 0, 200, false, 0);
 		for(unsigned i = 0; i < count; i++)
@@ -562,16 +595,9 @@ void GOSettings::Flush()
 		cfg.WriteString(wxT("AudioGroups"), wxString::Format(wxT("Name%03d"), i + 1), m_AudioGroups[i]);
 	cfg.WriteInteger(wxT("AudioGroups"), wxT("Count"), m_AudioGroups.size());
 
-	const wxString SOUND_PORTS = wxT("SoundPorts");
-	const wxString ENABLED = wxT(".Enabled");
-
-	for (const wxString &portName: GOSoundPort::getPortNames())
-	{
-	  cfg.WriteBoolean(SOUND_PORTS, portName + ENABLED, m_PortsConfig.IsConfigEnabled(portName));
-	  const wxString prefix = portName + ".";
-	  for (const wxString &apiName: GOSoundPort::getApiNames(portName))
-	    cfg.WriteBoolean(SOUND_PORTS, prefix + apiName + ENABLED, m_PortsConfig.IsConfigEnabled(portName, apiName));
-	}
+	save_ports_config(
+	  cfg, SOUND_PORTS, GOSoundPortFactory::getInstance(), m_SoundPortsConfig
+	);
 
 	for(unsigned i = 0; i < m_AudioDeviceConfig.size(); i++)
 	{
