+++ src/grandorgue/sound/ports/GOSoundPort.h
@@ -7,7 +7,6 @@
 #ifndef GOSOUNDPORT_H
 #define GOSOUNDPORT_H
 
-#include "sound/GOSoundDevInfo.h"
 #include "settings/GOPortsConfig.h"
 
 #include <wx/string.h>
@@ -18,8 +17,6 @@ class GOSound;
 class GOSoundPort
 {
 protected:
-  static const std::vector<wxString> c_NoApis; // empty Api list for subsystems not supporting them
-  
   GOSound* m_Sound;
   unsigned m_Index;
   bool m_IsOpen;
@@ -33,32 +30,7 @@ protected:
   void SetActualLatency(double latency);
   bool AudioCallback(float* outputBuffer, unsigned int nFrames);
 
-  static wxString composeDeviceName(
-    wxString const &subsysName,
-    wxString const &apiName,
-    wxString const &devName
-  );
-
 public:
-  class NameParser
-  {
-  private:
-    const wxString &m_Name;
-    size_t m_Pos;
-
-  public:
-    NameParser(const wxString &name): m_Name(name), m_Pos(name.IsEmpty() ? wxString::npos : 0) { }
-    NameParser(const NameParser &src): m_Name(src.m_Name), m_Pos(src.m_Pos) { }
-
-    const wxString &GetOrigName() const { return m_Name; }
-    bool hasMore() const { return m_Pos != wxString::npos; }
-
-    const wxString GetRestName() const 
-    { return hasMore() ? m_Name.substr(m_Pos) : wxT(""); }
-
-    wxString nextComp();
-  };
-  
   GOSoundPort(GOSound* sound, wxString name);
   virtual ~GOSoundPort();
 
@@ -68,14 +40,6 @@ public:
   virtual void Close() = 0;
 
   const wxString& GetName();
-
-  static const std::vector<wxString> & getPortNames();
-  static const std::vector<wxString> & getApiNames(const wxString & portName);
-  
-  static std::vector<GOSoundDevInfo> getDeviceList(const GOPortsConfig &portsConfig);
-  static GOSoundPort* create(const GOPortsConfig &portsConfig, GOSound* sound, wxString name);
-
-  static void terminate();
 
   wxString getPortState();
 };
